requirements.txt
Django~=5.1.3
djangorestframework~=3.15.2
django-filter~=25.1
django-jazzmin==3.0.1
djoser==2.3.1
drf-yasg==1.21.8
django-cleanup==9.0.0
django-debug-toolbar==5.0.1
django-import-export==4.3.4
djangorestframework==3.15.2
pillow==11.0.0
django-jazzmin==3.0.1
